+++ CHANGELOG.md
@@ -17,6 +17,7 @@ See the [v1 milestone for more details](https://github.com/DubPlus/DubPlus/miles
 - "Auto AFK", when user is no longer focused on the page it starts a timer that auto-enables AFK when it expires [#38](https://github.com/DubPlus/DubPlus/issues/38)
 - Snooze Video - hide the video for 1 song [#118](https://github.com/DubPlus/DubPlus/issues/118)
 - Grab Response - send a message to the chat when you grab a song [#107](https://github.com/DubPlus/DubPlus/issues/107)
+- Collapsible Images - images in chat can be collapsed/expanded [#74](https://github.com/DubPlus/DubPlus/issues/74)
 - Modal improvments [#123](https://github.com/DubPlus/DubPlus/issues/123)
   - show default value within the modal
   - disable options if user leaves modal blank and there is no default value